+++ Order-Core/src/main/java/com/goormgb/be/ordercore/club/dto/response/ClubDetailGetResponse.java
@@ -2,28 +2,28 @@ package com.goormgb.be.ordercore.club.dto.response;
 
 import java.math.BigDecimal;
 
-import com.goormgb.be.ordercore.club.entity.Club;
-import com.goormgb.be.ordercore.stadium.entity.Stadium;
-import com.goormgb.be.ordercore.state.entity.TeamSeasonStats;
+import com.goormgb.be.domain.club.entity.Club;
+import com.goormgb.be.domain.entity.Stadium;
+import com.goormgb.be.domain.state.entity.TeamSeasonStats;
 
 public record ClubDetailGetResponse(
-		Long clubId,
-		String koName,
-		String logoImg,
-		String clubColor,
-		StadiumDto stadium,
-		String homepageRedirectUrl,
-		CurrentSeasonStatsDto currentSeasonStats
+	Long clubId,
+	String koName,
+	String logoImg,
+	String clubColor,
+	StadiumDto stadium,
+	String homepageRedirectUrl,
+	CurrentSeasonStatsDto currentSeasonStats
 ) {
 	public static ClubDetailGetResponse of(Club club, TeamSeasonStats stats) {
 		return new ClubDetailGetResponse(
-				club.getId(),
-				club.getKoName(),
-				club.getLogoImg(),
-				club.getClubColor(),
-				StadiumDto.from(club.getStadium()),
-				club.getHomepageRedirectUrl(),
-				stats == null ? null : CurrentSeasonStatsDto.from(stats)
+			club.getId(),
+			club.getKoName(),
+			club.getLogoImg(),
+			club.getClubColor(),
+			StadiumDto.from(club.getStadium()),
+			club.getHomepageRedirectUrl(),
+			stats == null ? null : CurrentSeasonStatsDto.from(stats)
 		);
 	}
 
@@ -34,27 +34,27 @@ public record ClubDetailGetResponse(
 	}
 
 	public record CurrentSeasonStatsDto(
-			int seasonYear,
-			Integer seasonRanking,
-			Integer wins,
-			Integer draws,
-			Integer losses,
-			BigDecimal winRate,
-			BigDecimal battingAverage,
-			BigDecimal era,
-			BigDecimal gamesBehind
+		int seasonYear,
+		Integer seasonRanking,
+		Integer wins,
+		Integer draws,
+		Integer losses,
+		BigDecimal winRate,
+		BigDecimal battingAverage,
+		BigDecimal era,
+		BigDecimal gamesBehind
 	) {
 		public static CurrentSeasonStatsDto from(TeamSeasonStats stats) {
 			return new CurrentSeasonStatsDto(
-					stats.getSeasonYear(),
-					stats.getSeasonRanking(),
-					stats.getWins(),
-					stats.getDraws(),
-					stats.getLosses(),
-					stats.getWinRate(),
-					stats.getBattingAverage(),
-					stats.getEra(),
-					stats.getGamesBehind()
+				stats.getSeasonYear(),
+				stats.getSeasonRanking(),
+				stats.getWins(),
+				stats.getDraws(),
+				stats.getLosses(),
+				stats.getWinRate(),
+				stats.getBattingAverage(),
+				stats.getEra(),
+				stats.getGamesBehind()
 			);
 		}
 	}
